Update core edit_* tables with various fixes
- import script with write to file option
- fix file names for functions
- add generic (non edit tables) for only set date, set uid or combine
- fix edit table edit_language insert data
- all trigger create remove the drop on exists, as in the flow the trigger will never exists

+++ 4dev/database/bin/import_data.sh
@@ -6,20 +6,25 @@
 #exit;
 
 # if flagged 'y' then it will ask after each import to continue
-development='y';
+development='n';
+# do not import anything, just order flow output
 test='n';
-input='';
+# write to file do not write to DB directly
+write='y';
 # database connection info
 db='<db name>';
 host='<db host>';
 user='<db user>';
-schema="public";
+schemas='public';
 export PGPASSWORD='';
-
 # log files
-error_file="log/error";
-output_file="log/output";
+error_file='log/error';
+output_file='log/output';
+data_file='database_create_data.sql';
 
+if [ "$write" = 'y' ]; then
+	rm -f "${data_file}";
+fi;
 if [ ! -f ORDER ]; then
 	echo "Could not find ORDER file";
 	exit;
@@ -40,26 +45,37 @@ while read file <&3; do
 	if [ -f "$file" ]; then
 		for path in "$schemas"; do
 			echo "[+] WORK ON '${file}' @ '${path}'";
+			# skip all on test
 			if [ "$test" = 'n' ]; then
-				echo "=== START [$file] ===>" >> ${error_file};
-				psql -U ${user} -h ${host} -f "${file}" ${db} 1>> ${output_file} 2>> ${error_file}
-				echo "=== END   [$file] ===>" >> ${error_file};
-			fi;
-			if [ "$development" = "y" ]; then
-				echo "Press 'y' to move to next. Press 'r' to reload last file. ^c to abort";
-			fi;
-			while [ "$development" = "y" ] && [ "$input" != "y" ]; do
-				read -ep "Continue (y|r|^c): " input;
-				if [ "$input" = "r" ]; then
-					echo "Reload File '${file}' ...";
-					if [ "$test" = 'n' ]; then
-						echo "=== START RELOAD [$file] ===>" >> ${error_file};
-						psql -U ${user} -h ${host} -f "${file}" ${db} 1>> ${output_file} 2>> ${error_file}
-						echo "=== END RELOAD   [$file] ===>" >> ${error_file};
+				# write to file
+				if [ "$write" = 'y' ]; then
+					echo "-- START: ${file}" >> ${data_file};
+					cat "${file}" >> ${data_file};
+					echo "-- END: ${file}" >> ${data_file};
+				else
+					# write to DB
+					echo "=== START [$file] ===>" >> ${error_file};
+					psql -U ${user} -h ${host} -f "${file}" ${db} 1>> ${output_file} 2>> ${error_file}
+					echo "=== END   [$file] ===>" >> ${error_file};
+					# next wait for dev
+					if [ "$development" = "y" ]; then
+						echo "Press 'y' to move to next. Press 'r' to reload last file. ^c to abort";
 					fi;
+					# loop run for reload on failed
+					while [ "$development" = "y" ] && [ "$input" != "y" ]; do
+						read -ep "Continue (y|r|^c): " input;
+						if [ "$input" = "r" ]; then
+							echo "Reload File '${file}' ...";
+							if [ "$test" = 'n' ]; then
+								echo "=== START RELOAD [$file] ===>" >> ${error_file};
+								psql -U ${user} -h ${host} -f "${file}" ${db} 1>> ${output_file} 2>> ${error_file}
+								echo "=== END RELOAD   [$file] ===>" >> ${error_file};
+							fi;
+						fi;
+					done;
+					input='';
 				fi;
-			done;
-			input='';
+			fi;
 		done;
 	elif [[ ${file::1} != "#" ]]; then
 		echo "[!] COULD NOT FIND FILE: '${file}'";
